+++ src/steps/user-group/index.ts
@@ -1,10 +1,18 @@
 import {
+  createDirectRelationship,
   Entity,
   IntegrationStep,
   IntegrationStepExecutionContext,
+  Relationship,
+  RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
 
-import { Entities, Steps } from '../constants';
+import {
+  ACCOUNT_ENTITY_KEY,
+  Entities,
+  Relationships,
+  Steps,
+} from '../constants';
 import { createUserGroupEntity } from './converter';
 import { createSonarqubeClient } from '../../provider';
 import { SonarqubeIntegrationConfig } from '../../types';
@@ -13,13 +21,24 @@ export async function fetchUserGroups({
   instance,
   jobState,
 }: IntegrationStepExecutionContext<SonarqubeIntegrationConfig>) {
+  const accountEntity = (await jobState.getData(ACCOUNT_ENTITY_KEY)) as Entity;
   const client = createSonarqubeClient(instance.config);
 
   const convertedUserGroups: Entity[] = [];
+  const relationships: Relationship[] = [];
   await client.iterateUserGroups((userGroup) => {
-    convertedUserGroups.push(createUserGroupEntity(userGroup));
+    const userGroupEntity = createUserGroupEntity(userGroup);
+    convertedUserGroups.push(userGroupEntity);
+    relationships.push(
+      createDirectRelationship({
+        _class: RelationshipClass.HAS,
+        from: accountEntity,
+        to: userGroupEntity,
+      }),
+    );
   });
   await jobState.addEntities(convertedUserGroups);
+  await jobState.addRelationships(relationships);
 }
 
 export const userGroupSteps: IntegrationStep<SonarqubeIntegrationConfig>[] = [
@@ -28,6 +47,7 @@ export const userGroupSteps: IntegrationStep<SonarqubeIntegrationConfig>[] = [
     name: 'User Groups',
     entities: [Entities.USER_GROUP],
     executionHandler: fetchUserGroups,
-    relationships: [],
+    relationships: [Relationships.ACCOUNT_HAS_USER_GROUP],
+    dependsOn: [Steps.ACCOUNT],
   },
 ];